Initial Commit
Another thing I've been trying to figure out for a while, is menu declarations, whether it should be component and or HTML-driven, or if it should purely be a JS-defined thing.

Either way, if I were to hypothetically use a builder function (seems like a neat way to apply for this scenario), it seems to nicely chain things together.

Today's discovery was realizing how to strictly-type accelerator keyboard combinations. I very much enjoy this. There might be a better way to type this out, I'm not totally sure. This works great already though.

I think part of what I'm curious about, is to define the menu using state-like objects, rather than element components themselves. This seems to be a great breakthrough I think, as then you aren't coupling it to the UI rendering itself, something I have struggled with in terms of conceptualization, and in practice.

I think we're landing so I gotta go!