Menu Systems Module Move
Early Porcupine Tree definitely sounds like Orbit (I'm listening to Orbit, so the way I recognized it was the other way around, realizing how Up the Downstair sounds like Orbit).

Made a cool logo for my music, this time I am going to scan the paper, and make a vector out of it, rather than draw it with paint tools in Photopea. I want the kind of off-kilter drawing shapes of it. It's shaped nicely, but not perfect per se. I like that about when you draw and create your own art.