Menu Systems Module Rename
Going to start implementing some of the functions, and in doing that, I can't use fully ambient modules for things.

I do love TS for that, that you can demo out the shapes of the API, without having to implement the internal code to handle that itself. You could do this with C in theory too, I guess Rust in a way, but the compiler will get mad at you for it, and prevent you from compiling. Probably the same with GCC, so TS is kind of cool and unique in that way.

Just finished Orbit Orbit. I love the songs on there, but man that is not a nice album listen. That's my weighted opinion of course. It just feels very taxing. Maybe because as a listener I very much enjoy variety. It's okay for songs to hone in on a single style, I feel the album shouldn't do this though. Again, if it was the goal, then it's just not my thing. Because if it is meant to keep you locked in in that way, it works very well. For me, it kind of makes me uneasy a little, to know what comes next, and to realize how much of the album is still left to listen to.