Acc Sorting
Realized with this addition, that there's a set of duplicates which I hadn't seen yet. I think it's because TS handles this for us at the type level, hence why I missed it.

I absolutely love the mix of this song, seriously. I can't remember if I had already finished Recreational Outrage, but the sound of it definitely reminds me of that time. And if it didn't influence me for that, it is definitely influencing me now. It's so dynamic yet clear. I fucking love it, for lack of better words 😂.