Numeric Key Strings
This is more maintainable in terms of types and things like that.

Christeen + 4 Demos

I love this second half to Infinity.

Even with these earbuds, it sounds freaking awesome. And it is the lossless, non-remastered version. I prefer the original to the 2023 remaster. I like the edginess of it. I think someone else mentioned that online somewhere too, that the grittiness is what makes the album what it is, and after becoming aware of it, I feel so too.

It looks so cool to hear 'star child rise' and to look out the plane window and see pink sunset clouds. Amazing, and fully greatful for everything I have in my life.