Accelerator State
Parsing the accelerator strings to be accessible as certain properties and things like that. This was a big challenge in STE, and figuring out how to do this is a really good thing. Let alone that it's strongly-typed as well, which is great. Glad I made the constant values too, because that makes it easier to check whether something is valid or not, as they are simply just a real list of values.

The other half to this is I need to be able to parse KeyboardEvent key info into this same kind of AcceleratorState object, then I can compare the two to find out what they are.

After having said that, I wonder if actyally instead it might make sense to just serialize the KeyboardEvent key info down to a formatted accelerator string instead, too? That makes it extremely trivial to validate. I will look into that. It probably makes more sense to do that. It is easier to format, and less conversions all over the place too.