MetaKey to SuperKey
Realized I had mixed up the names of these two. Meta key is the specific key which is either the Win or Cmd button. Super is the platform-agnostic control key, like Ctrl or Cmd. So while they kind of have crossover in what they are talking about, their intent is focused on two different things.

For the accelerator strings, I don't want to worry about the platform, just simply to handle the literal intent. So maybe I will add a specific MetaKey handler in the literal sense, for the Win or Cmd key, but I don't know if it really makes sense, because only macOS is able to detect it's keypresses. I don't think Windows lets you intercept the Meta key.

Note that I do still need to detect the Meta key, it is crucial to this implementation. It's just that the accelerators themselves should be focused on either Super, Cmd, or Ctrl, rather than Meta itself.