Apple Device Constant
...rather than a function call.

Not totally on board with the new Apple updates. More for the bugs and inconsistencies than the design changes in what they are trying to achieve. I do like the neumorphism concept, but I don't like how it's executed. It feels a bit cluttered and or distracting, rather than something natural. I think it will get better with time though, and I want to start dealing with it now anyways, since the damage has been done; I don't want to be worrying about trying to manage old OS usage, I'd rather try to put that energy into using the new OS, even if I don't like it. There's parts of my workflow that it changed, and things my eyes are used to looking at in terms of common usage patterns, things like that. But again, I can learn to change those, it's just that I got used to them. It's okay! I just have to try. I am glad I did, even if it is still urking me to use it.

Going to start building an ID3 tag editing tool, and call it urit. As in, *tag, 'urit'!*

SYL City non-remastered is brutally awesome, so glad to have it instead of the remaster. It is in lossless too. Also got CAB 1 and CAB 2. All of these are some of my favorite albums ever. All of them are priceless in terms of my music taste and what I love listening to. Lossless only makes that even better. I love music!