Accelerator Checker
Realized right as I finished this that I really don't need to do the checks like this. There is a deterministic list of all of the valid shortcuts anyways, so I should just check that it's part of that list. Yeah!

I could just concatenate the string parts based on what is present in the event, then validate that what it made was one of the ones available as something to be assigned to. And it's all type-driven too, would be extra awesome! Would be much simpler than this code here too, as this code is much more complex than simply validating against the data we already constructed in the other file.

Wish I could go for another run around Boston, it was so fun! I love how walkable it is over there. Have to drive more over here for sure. But we do have beautiful mountains too though, so that's a trade-off as well.

Yeah, I am not minding macOS changing, quite as much. Or rather, I am trying to tell myself that just so I stop worrying about it. I can still get stuff done, as you can see right now. Just need to not worry that my usage is changing because muscle memory and where I used to look for things is now elsewhere.