Accelerator Checker from List
Listening to Above the Roses, I love this album! Marco got me into the self-released scene I think, it has been so inspiring to see what he's done on his own. That's what got me into doing music on my own too. I have to lend it to him specifically for that. Makes me want to make more again. I love music!

"Daddy Won't Be Home Tonight" is such a Marco song, it's awesome. I love it so much. Not sure what you'd call that instrumental break where everything is layered on top of each other, but that thing is so cool. I want to learn how to compose like that. I think I get locked into improv and I end up forgetting that I wanted to try that out. Will try keeping it in the front of my mind, because I think it has to come naturally for me to feel like it works, in terms of my own music.

I am hearing a bunch of my influences in my own music too, I think parts of this song showed up in Haunted in a way, like the double drum part in this song, as well as the instrumental line-up thing. I don't think I used that specifically, but something about it seemed to backreference this, but I don't remember how exactly. Same with Elkosaurus and the Schattenspiel keyboard aesthetic. In my head there is a connection, but I don't remember if the resulting project conveyed it to that extent or not.

It's awesome just how much music brings you back to a certain time period. This album specifically reminds me of "senior year", when I was at home doing work, during the last phases of the pandemic. Man, what a hard, difficult time it was. I was just writing in my notebook a few days ago how recently it's felt harder than that point. I think it was a longer period of time where things felt bleak like that back then, and it was even stronger as of recently, but not as long in duration.

I am super curious about the mix on some songs, now that I think about it.

Marco seems to place his vocals way up in the mix, as well as the drums. Well, everything seems really balanced and I really like that. He uses a lot of panning too, which I also enjoy. CAB seems to actually place the guitar pretty far down in the mix. Strapping doesn't have much dynamics. Marco's vocals (on Roses at least) seems to not have a de-esser. And I don't think he used much compression or EQ, or at least there seems to be a lot of dynamics I think. Whether that's good or bad, that's up to interpretation. I very much enjoy how these aspects cater to the mix. Again, it could use a little bit of boots in some aspects in the mix phase I think, but I do value how clear the mix is too, in that it just is very breathful in a way. I felt his music has a lot of clarity in it, and I think it is because it's so raw in a way, both lyrically, emotionally, stylistically, it's beautiful. He makes what he wants to, and he's not afraid to put the vocals up in front of the mix. I am curious if anyone mixed or mastered it, or if he did it at home. *just went and looked it up, looks like he did it all at home himself! In that case, huge kudos! It sounds great! Much like myself, with the nicest of observation, I think there are only a few things that could have helped the mix, like the de-esser and small things like that. Other than that, I love the sound. If anything, back before I knew about any of this audio stuff, I think I digged the raw sound more than polished, which I also thank him for. It's a great headphone listen, maybe not as consistent for listening in the car though, as it's hard to hear everything all at once, if that makes sense. I feel the same about most of my songs too. I am trying to work on it. It's a hard balance to figure out what a mix needs, and even more importantly, doesn't need.